Chat2Anything-admin/dockerdata/requirements.txt
Flask
Flask-APScheduler
Flask-Migrate
Flask-Login
Flask_Reuploaded
Flask-SQLAlchemy
flask-marshmallow
marshmallow-sqlalchemy
PyMySQL
psutil
Flask-Mail
Pillow
validators
cryptography